Fix warnings
git-svn-id: https://svn.isis.vanderbilt.edu/GMESRC/trunk@1599 b35830b6-564f-0410-ac5f-b387e50cb686

+++ GME/XmlBackEnd/DOMErrorPrinter.cpp
@@ -36,7 +36,7 @@ bool DOMErrorPrinter::handleError(const DOMError &domError)
 		memset( line_nmbs, 0, sizeof(line_nmbs));
 
 		smart_Ch f_str = XMLString::transcode(loc->getURI());
-		sprintf( line_nmbs, "Line: %u Column: %u of File: ", loc->getLineNumber(), loc->getColumnNumber());
+		sprintf( line_nmbs, "Line: %u Column: %u of File: ", (unsigned)loc->getLineNumber(), (unsigned)loc->getColumnNumber());
 		std::string linfo = std::string( line_nmbs) + f_str;
 
 		// send msg
